Changed raw_product to be the .area feature of the buffer used for the compressed result, instead of the buffer itself (SPECIAL instead of TO_SPECIAL).
After noticing that with small target strings, the compress
routine can sometimes return a buffer-too-small error, I extended the
`compress_target' routine to check for this error and try again with a
larger buffer.  In the case where I encountered this problem (the target size
was 5 and original buffer size was 31), the 2nd call succeeded.

Added a couple print statements to time the compression.

Changed the invariant to include -1, which is the default compression level
value.