New classes and changes to support the new "add user" request.  (User will be used for logins, session tracking, and related functionality.)
New classes and changes to support the new requirement that the "add
person" request returns the ID of the new person.  Made some infrastructure
changes and added some new classes to support this, including a SQL_COMMAND
hierarchy and use of the EVENT_SUPPLIER class (which needs to be moved into
the eiffel_library) to allow first obtaining an ID for the new PERSON
(obtained as the next value in a sequence), placing that ID into the
"insert person" SQL command, and obtaining the ID value (via the
EVENT_SUPPLIER callback mechanism) and sending it back to the client.

Other small improvements.