Redefined initialize to initialize its 'indexable' field (if it is of the right type).  This allows the index extractor to be used in a more simple context than before.  For example, the parabolic-sar macl script (created before this change) uses a unary_linear_operator to hold the index_extractor, which, in turn, contains the same unary_linear_operator object (as its 'indexable') from which it extracts the current index.  With the change, the top unary_linear_operator is no longer needed: The index_extractor can be used directly and it can use a unary_linear_operator, or simply a basic_linear_command, as its indexable.
However, this change created a problem with existing cyclic structures,
such as the unary_linear_operator -> index_extractor -> unary_linear_operator
one described above.  To fix this, I added an 'initialized' boolean
attribute to the class and used it in the initialize routine to prevent
initializing the 'indexable' more than once in one "initialization
session".  Tested by hand and with the server/test/run_tests script with
success.