Bug fix in 'suppliers':  Just as happened with 'descendants', an infinite loop would occur when there is a cycle in the command tree.  Fixed by using the same locking scheme as with 'descendants'.  As with 'descendants', the only class whose instances can validly be part of a cycle is index_extractor, so the same implementation as with 'lock_descendants', etc. was provided for 'lock_suppliers', etc.
index_extractor: Changed the "state-index" constants into ordinary
functions.  This appears to fix the problem of causing an incompatible
indicators_persist file when a new "state" is added.

command: Fixed two bugs in suppliers: It was iterating over 'descendants'
and calling 'suppliers' on each descendant, causing 'suppliers' to be
called, unecessarily, more than once on an object.  Changed to iterate
over 'children' instead.  And the line

 				Result := descendants.first.suppliers

(which after the first bug fix is:

 				Result := children.first.suppliers
)

But it needs to be:

 					Result := l.first.suppliers

since l, which was assigned to 'children' is to be used to iterate over
each child and obtain its 'suppliers'.


These fixes were all moderately tested, with success.