Cleanup of code marked with !! comments.  Consisted of removing replacement code or old obsolete code, or comments reminding to check if a new piece of code is correct.
These comments were inserted quite a while ago (several months) and most
likely the replacement code or the code to test has been executed many
times with no errors, implying that the code is correct.  It is likely that
this is the case, but if a subtle bug appears that is hard to trace, try
using this cvs log entry as a clue indicating possible places in the code
that could be causes of the bug.  (Do a cvs diff, of course, to see what
was changed.)