Fix in 'field_count': Was returning an incorrect result when called with 'field_index' > 1.
The fix is mainly in this section:

			if not s.is_empty then
				su.set_target (s)
				fcount := su.tokens (field_separator).count
				if field_index = 0 or else fcount = field_index then
					Result := fcount
				else
					Result := fcount + field_index - 1
				end
			end


Without assertions, 'field_count' will not be called often, so its
calculation will not be a major performance hit, and it will only be
called when in a state where the above change is not needed (or was not
needed in the previous version).  But the fix is needed, of course for
use in debugging-print statements and with assertions on.

Inserted some debugging code, which will come out soon.

Fixed a bug that manifested after recent changes in other files -
partly documented in the 'task 43' entry: Date/time: 2004-12-12, 20:7:

'after_last_record' in INPUT_FILE was false at the end of input instead of
true because 'end_of_file' was returning false.  I'm not sure what changed
to cause this - probably a subtle change re. reading fields or records
(with the fix of the above bug) that caused the file not to be read beyond
the last record.  Fixed by copying the lines:

            read_character
            if not after then
                back
            end

from the end of 'discard_current_record' to the end of
'advance_to_next_record'.

The current fix is now a little more complex:

			if not end_of_file then
				check
					is_readable: readable
				end
				read_character
				check
					not_closed_or_empty: count > 0 and not is_closed
				end
				if not end_of_file then
					back
				else
					after_last_record := True
				end
			else
				after_last_record := True
			end

I think this can be put into a separate routine and called by both
'advance_to_next_record' and 'discard_current_record'.

Also, 'after_last_record' was changed to an attribute and code that uses
it changed accordingly - for performance optimization.

And, another bug fix - 'position_cursor' needs to set the field and record
indexes and 'after_last_record'.