Merged estudio-5.6 branch change ( svn merge -r 587:588 file:///home/svn/eiffel_library/branches/estudio-5.6 ) onto the trunk.
Tested the affected functionality ('view an indicator') successfully.  The
change fixed a bug resulting from the port to ES 5.6.  Here is the log
entry for the original change in the branch:

Fixed a bug resulting from the modifications to compile with ES 5.6: TUPLE
no longer has a 'force' feature - changed to use 'put' instead.  But the
result was "putting" into an empty tuple, which violates the 'put'
precondition.  There appears to be no way to create a tuple and end up with
it being non-empty, except indirectly - using the [element, ...] construct.
And it looks like this was the way it should have been used in the first
place - i.e., use [Current] as the argument to
report_function.set_operands, rather than creating a tuple explicilty and
using it as the argument.

Tested successfully.  The bug appears when, from the command line, the
'view and indicator' command is used.  The resulting segv no longer occurs
and the CLI behaves correctly.