Merged estudio-5.6 branch change ( svn merge -r 591:592  file:///home/svn/eiffel_library/branches/estudio-5.6 ) onto the trunk.
Compiles fine - not tested, but the change is small and is not likely to
introduce bugs into this estudio 5.1 version.

Log file from the original checkin:

Fix for a bug I encountered while investigating the STORABLE compiler bug -
If I remember right, an invariant was violated with respect to the
field_separator feature (not void or not empty, I think).

Also inserted a debug/end block to get rid of an old debugging print cmd.