Changes to bring old code in compliance with modern ISE compiler
A large number of changes to get this old code to compile with the new ISE
compiler. Changes consisted mostly of removing the ' is' in
procedure/function definitions, changing "is" to "=" for constant
definitions, and changing the old 'indexing' keyword at the beginning of a
class to the new 'note' keyword.  The application (market analysis system)
does not yet compile all the way, but I decided to commit the current
changes, since they were pretty extensive in terms of # of files changed.
The compilation currently gets to degree 3 (checking types - It starts at
degree 6).

Had to make some changes due to the new compiler being pickier than the old
version - such as some undefines and renames, and variable type changes
(e.g., ARRAY[STRING] -> LIST[STRING]).  There's quite a few more degree-3
fixes to go.