14.05 port - Extensive changes/fixes for correct basic functionality
Now that the code (for MAS) compiles, the rather large task of getting it to
run correctly has begun. Things are different with the new compiler - some
of the "type casts" (via if attached ...) were the wrong solution (and at
least one existing type cast - via ?= - was wrong), with the fix in (at
least) some of those cases to get rid of the type casts and employ, in
TREE_NODE, effective anchoring with "like ..." in the right places (along
with related type changes in some descendants). (See, for example, the
*tuple_printer* classes and MARKET_FUNCTION.)

The finalized executable (with and without assertions) will now correctly
(via the command-line interface) read a stock-data file, print the stock
data, and process and print the data filtered through an indicator/function
(such as moving average, stochastic, etc.).  It also runs this scenario
successfully in estudio if assertions are turned off.  (Running with them
on triggers an invariant violation: input_target_relation in
TWO_VARIABLE_FUNCTION that appears to not make sense.)

A list of some (but not all - not enough time) other changes:

  - Added a GENERIC_FACTORY (which inherits from FACTORY) class (with a
    generic attribute - product: G), which most of the (mostly MAS-related)
    classes that used to inherit from FACTORY now inherit from.
    (MARKET_EVENT_PARSER still inherits directly from FACTORY.)  FACTORY is
    no longer generic (same as in the old version of the code).

  - Corrected a couple of the 'across..loop ... if attached...' blocks that
    I had added to use the item feature (e.g., "if attached...cursor.item
       as f" instead of "if attached...cursor as f" in
    MARKET_EVENT_GENERATOR.)

  - Had to make a few changes related to the fact that an ARRAYED_LIST is
    no longer an ARRAY, some of which involved simply changing the type of
    a function, attribute, or variable from ARRAY to ARRAYED_LIST (and
    calling make_empty instead of make(startidx, endidx).

  - Changed the default mailer from elm to mutt.

  - STOCK_SPLIT_SEQUENCE was very likely incorrectly converted; this has
    been fixed - probably.  Test/verify/...fix-again[if-necessary]/etc.

  - Corrected a wrong postcondition in TRADABLE_LIST.

  - A couple calls to the make(minidx, maxidx) creation routine in ARRAY
    were changed to make_empty instead due to an apparent bug that surfaces
    when running in estudio (not, I believe, in the compiled/finalized
    executable), along with associated changes of 'put' to 'force' to make
    up for the array not having been "presized" on creation.

  - In the eiffel_library, DATA_SCANNER no longer inherits from FACTORY.

  - In REAL_SETTER, stream.read_real -> stream.read_double - Was needed as
    a result of the change throughout to use DOUBLE instead of REAL (which
    change seems wise re. accuracy now that 64-bit machines are common).

  - Made some fixes in the command-line parsing functionality - Now options
    are parsed and handled correctly, just as with the old code with old
    compiler.