Further implementation of socket-enhancements
EIFFEL_LIBRARY:

CONNECTED_SOCKET_POLL_COMMAND:  It now depends on CONNECTED_SOCKET_PROCESSOR,
which means it needs to be moved to where CONNECTED_SOCKET_PROCESSOR
resides - the mas/finance repository.  (This could probably be fixed via
some refactoring and a new class, but there may not be time for that.)
Added some 'cleanup' logic (remove_read_command, close socket) via new
feature, final_cleanup.

general: Added some error-handling/reporting logic.

MAS/FINANCE:

general:

set CONNECTION_CACHE size to 1 to allow testing of the
remove-the-oldest-poll-command behavior (and resulting cleanup) with only 2
clients running at the same time.

Some changes to return the server-will-not-close (after responding)
response codes for error conditions to let the client know that (as with a
normal/ok response) it should not close its connection - i.e., it should
use the same socket connection for the next request rather than creating a
new one.

Added some logic to ensure that when a CONNECTED_SOCKET_POLL_COMMAND is
removed from the CONNECTION_CACHE the following behavior occurs:

  - The poll command and its associated socket are available for one more
    request/response to the associated client.
  - When the response is sent, the connection-will-be-closed response code
    is sent (rather than the usual ...will-remain-open code) to reflect the
    fact that the associated socket has been kicked out of the cache.
  - After the response is sent, the associated socket is closed.

This design - of allowing one more request/response communication with the
client to occur after removal from the cache - appears to be the most
straightforward way to ensure that the client is notified that the socket
will be closed after the last response from the doomed socket.