Added open_for_reading "require" to some methods.
In INPUT_RECORD_SEQUENCE 'open_for_reading' is always true.  One
descendant, INPUT_MEDIUM (and its descendant, obviously),
redefines it to is_open_read (from IO_MEDIUM). This ensures that
the methods with the new precondition, such as 'field_count',
don't cause problems, such as a core dump, when OS features such
as file access are invoked on a file that is not bound to an
actual file on the filesystem (i.e., is not open).

.gitignore: More "garbage" filename patterns added.