Added concept of "parent" to TREE_NODEs
Changes to allow the MAS server to support potential implementation
of uniqueness of function-parameter names by adding a "parent" query
for the TRADABLE_FUNCTIONs, COMMANDs, etc. that changeable
parameters are potentially associated with - so that the "parameter"
can use this "parent" knowledge of its associated object (or ask the
object itself to get the information) [and, by association, parent-of-
parent, etc. (thus the additional new "ancestors" query)] to obtain
a "name" (or unique tag, or whatever one would like to call it) that
no other parameter of the target tree (which will be, i.e., either
an "indicator" or a "market-event generator") will have.  This
"parent" query, and supporting features, has been added to
TREE_NODEs, which is the ancestor of the classes in question
(COMMANDs, etc.)  The "initialize_from_parent" procedure introduced
in TREE_NODE needed to be implemented by key descendent classes to
provide the needed "parent-setting" functionality.  Also, calls to
"initialize_from_parent" needed to be added to appropriate places in
the code (such as "set_operand" methods) to provide parent-setting
at the right time (i.e., when the object is being contructed or
configured).

Also, added a full_name query to COMMAND, which might (or might
not) be used for the parameter-name uniqueness implementation.